+++ packages/verified-fetch/test/verified-fetch.spec.ts
@@ -14,6 +14,7 @@ import * as ipldJson from 'multiformats/codecs/json'
 import * as raw from 'multiformats/codecs/raw'
 import { identity } from 'multiformats/hashes/identity'
 import { sha256 } from 'multiformats/hashes/sha2'
+import pDefer from 'p-defer'
 import Sinon from 'sinon'
 import { stubInterface } from 'sinon-ts'
 import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
@@ -157,6 +158,7 @@ describe('@helia/verifed-fetch', () => {
       expect(ipfsResponse).to.be.ok()
       expect(ipfsResponse.status).to.equal(301)
       expect(ipfsResponse.headers.get('location')).to.equal(`ipfs://${res.cid}/foo/`)
+      expect(ipfsResponse.headers.get('X-Ipfs-Path')).to.equal(`/ipfs/${res.cid}/foo`)
       expect(ipfsResponse.url).to.equal(`ipfs://${res.cid}/foo`)
     })
 
@@ -836,6 +838,89 @@ describe('@helia/verifed-fetch', () => {
 
       const resp = await verifiedFetch.fetch(`http://example.com/ipfs/${cid}/foo/i-do-not-exist`)
       expect(resp.status).to.equal(404)
+    })
+
+    it('returns a 404 when walking dag-pb for non-existent path', async () => {
+      const fs = unixfs(helia)
+
+      const res = await last(fs.addAll([{
+        path: 'index.html',
+        content: Uint8Array.from([0x01, 0x02, 0x03])
+      }], {
+        wrapWithDirectory: true
+      }))
+
+      if (res == null) {
+        throw new Error('Import failed')
+      }
+
+      const resp = await verifiedFetch.fetch(`ipfs://${res.cid}/does/not/exist`)
+      expect(resp.status).to.equal(404)
+    })
+  })
+
+  describe('sessions', () => {
+    let helia: Helia
+    let verifiedFetch: VerifiedFetch
+
+    beforeEach(async () => {
+      helia = await createHelia()
+      verifiedFetch = new VerifiedFetch({
+        helia
+      })
+    })
+
+    afterEach(async () => {
+      await stop(helia, verifiedFetch)
+    })
+
+    it('should use sessions', async () => {
+      const getSpy = Sinon.spy(helia.blockstore, 'get')
+      const deferred = pDefer()
+      const controller = new AbortController()
+      const originalCreateSession = helia.blockstore.createSession.bind(helia.blockstore)
+
+      // blockstore.createSession is called, blockstore.get is not
+      helia.blockstore.createSession = Sinon.stub().callsFake((root, options) => {
+        deferred.resolve()
+        return originalCreateSession(root, options)
+      })
+
+      const p = verifiedFetch.fetch('http://example.com/ipfs/QmNnooDu7bfjPFoTZYxMNLWUQJyrVwtbZg5gBMjTezGAJA', {
+        signal: controller.signal
+      })
+
+      await deferred.promise
+
+      expect(getSpy.called).to.be.false()
+
+      controller.abort()
+      await expect(p).to.eventually.be.rejected()
+    })
+
+    it('should not use sessions when session option is false', async () => {
+      const sessionSpy = Sinon.spy(helia.blockstore, 'createSession')
+      const deferred = pDefer()
+      const controller = new AbortController()
+      const originalGet = helia.blockstore.get.bind(helia.blockstore)
+
+      // blockstore.get is called, blockstore.createSession is not
+      helia.blockstore.get = Sinon.stub().callsFake(async (cid, options) => {
+        deferred.resolve()
+        return originalGet(cid, options)
+      })
+
+      const p = verifiedFetch.fetch('http://example.com/ipfs/QmNnooDu7bfjPFoTZYxMNLWUQJyrVwtbZg5gBMjTezGAJN/foo/i-do-not-exist', {
+        signal: controller.signal,
+        session: false
+      })
+
+      await deferred.promise
+
+      expect(sessionSpy.called).to.be.false()
+
+      controller.abort()
+      await expect(p).to.eventually.be.rejected()
     })
   })
 })