+++ packages/interop/CHANGELOG.md
@@ -1,3 +1,47 @@
+## @helia/verified-fetch-interop [1.23.0](https://github.com/ipfs/helia-verified-fetch/compare/@helia/verified-fetch-interop-1.22.0...@helia/verified-fetch-interop-1.23.0) (2024-05-09)
+
+
+
+### Dependencies
+
+* **@helia/verified-fetch:** upgraded to 1.4.1
+
+## @helia/verified-fetch-interop [1.22.0](https://github.com/ipfs/helia-verified-fetch/compare/@helia/verified-fetch-interop-1.21.1...@helia/verified-fetch-interop-1.22.0) (2024-05-09)
+
+
+### Features
+
+* use blockstore sessions ([#50](https://github.com/ipfs/helia-verified-fetch/issues/50)) ([541dd64](https://github.com/ipfs/helia-verified-fetch/commit/541dd646b0e83b9c69ed32d7a29e964144ad03cf))
+
+
+
+### Dependencies
+
+* **@helia/verified-fetch:** upgraded to 1.4.0
+
+## @helia/verified-fetch-interop [1.21.1](https://github.com/ipfs/helia-verified-fetch/compare/@helia/verified-fetch-interop-1.21.0...@helia/verified-fetch-interop-1.21.1) (2024-05-08)
+
+
+### Bug Fixes
+
+* prevent interop timeouts with fast fixture loading ([#73](https://github.com/ipfs/helia-verified-fetch/issues/73)) ([a43d994](https://github.com/ipfs/helia-verified-fetch/commit/a43d9940cf56d36e4d348246730b144a59e2cc8a))
+
+## @helia/verified-fetch-interop [1.21.0](https://github.com/ipfs/helia-verified-fetch/compare/@helia/verified-fetch-interop-1.20.0...@helia/verified-fetch-interop-1.21.0) (2024-05-02)
+
+
+
+### Dependencies
+
+* **@helia/verified-fetch:** upgraded to 1.3.14
+
+## @helia/verified-fetch-interop [1.20.0](https://github.com/ipfs/helia-verified-fetch/compare/@helia/verified-fetch-interop-1.19.0...@helia/verified-fetch-interop-1.20.0) (2024-04-20)
+
+
+
+### Dependencies
+
+* **@helia/verified-fetch:** upgraded to 1.3.13
+
 ## @helia/verified-fetch-interop [1.19.0](https://github.com/ipfs/helia-verified-fetch/compare/@helia/verified-fetch-interop-1.18.0...@helia/verified-fetch-interop-1.19.0) (2024-04-15)